Implement API call for website.

+++ Core/Dto/Response/SearchMetadataDto.cs
@@ -2,6 +2,6 @@
 {
     public class SearchMetadataDto
     {
-        public long TotalPublishedNames { get; set; }
+        public int TotalPublishedNames { get; set; }
     }
 }
